pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "imageforge"
version = "0.1.0"
description = "Standalone local-first image generation service for eCardFactory"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0,<1.0.0",
  "uvicorn[standard]>=0.32.0,<1.0.0",
  "pydantic>=2.9.0,<3.0.0",
  "psycopg[binary]>=3.2.0,<4.0.0",
  "httpx>=0.27.0,<1.0.0",
  "pillow>=11.0.0,<12.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0,<9.0.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[tool.setuptools.packages.find]
include = ["app*"]
